replace infographicfilter to be infographicpanelrenderer that renders and filters panels
