+++ client/src/infographic/InfographicPanelRenderer.yaml
